+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 0.9.0
 
+## New Components
+
+* Multi Action Button 
+
 ## Selectable Table Rows
 
 * Table and TableAjax now have props of `selectable` and `highlightable`, enabling selectable or highlightable rows. Each event also emits events which can be used by developers with props of `onSelect` or `onHighlight`. Developers can also manually control the highlighting or selecting of rows using the same props on TableRow components.